command-line-tool: Build manifest in Guix inferior with channels.
* ravanan/command-line-tool.scm: Import (guix inferior).
(call-with-inferior): New function.
(run-command-line-tool): Accept channels argument and pass it on to
build-command-line-tool-script.
(build-command-line-tool-script): Accept channels argument and pass it
on to manifest-file->environment.
(manifest-file->environment): Accept channels argument. Build manifest
in Guix inferior when channels is provided.
* ravanan/workflow.scm (workflow-scheduler): Accept channels argument
and pass it on to run-command-line-tool.
(run-workflow): Accept channels argument and pass it on to
workflow-scheduler.
* bin/ravanan (main): Pass #f as channels to run-workflow.