guix: Work with a test suite interned to the store.
Instead of cwl-v1.2-conformance being a script that accepts a path to
the CWL v1.2 conformance test suite, we now work with a CWL v1.2 repo
origin object. This is more declarative, and there is no need for
manual cloning of the cwl-v1.2 git repo on the CI. The CI merely has
to run a script with no arguments. This provides more control here in
the ravanan repo and less configuration on the CI server.

We do likewise for the end-to-end tests. In this case, we work with
files from the ravanan repo, but have to first compile ccwl sources to
CWL workflows.